+++ pkg/host/serv.go
@@ -1,9 +1,9 @@
 package host
 
 type Serv interface {
-	Delete(id uint64) error
+	Delete(id uint32) error
 	GetAll() ([]*Host, error)
-	GetByID(id uint64) (*Host, error)
+	GetByID(id uint32) (*Host, error)
 	Store(u *Host) error
 	Update(u *Host) error
 }
@@ -18,11 +18,11 @@ func NewHostServ(repo Repo) Serv {
 	}
 }
 
-func (svc *hostServ) Delete(id uint64) error { return svc.repo.Delete(id) }
+func (svc *hostServ) Delete(id uint32) error { return svc.repo.Delete(id) }
 
 func (svc *hostServ) GetAll() ([]*Host, error) { return svc.repo.GetAll() }
 
-func (svc *hostServ) GetByID(id uint64) (*Host, error) { return svc.repo.GetByID(id) }
+func (svc *hostServ) GetByID(id uint32) (*Host, error) { return svc.repo.GetByID(id) }
 
 func (svc *hostServ) Store(u *Host) error { return svc.repo.Store(u) }
