+++ pkg/property/property.go
@@ -13,9 +13,9 @@ const (
 
 // Property model describes a single key value pair for a service(parameters). An example could be a port for HTTP checking
 type Property struct {
-	ID uint64 `json:"id,omitempty"`
+	ID uint32 `json:"id,omitempty"`
 
-	ServiceID uint64 `json:"service_id" gorm:"not null"`
+	ServiceID uint32 `json:"service_id" gorm:"not null"`
 
 	Key string `json:"key" gorm:"not null; default: null"`
