+++ pkg/service_group/repo.go
@@ -1,9 +1,9 @@
 package service_group
 
 type Repo interface {
-	Delete(id uint64) error
+	Delete(id uint32) error
 	GetAll() ([]*ServiceGroup, error)
-	GetByID(id uint64) (*ServiceGroup, error)
+	GetByID(id uint32) (*ServiceGroup, error)
 	Store(u *ServiceGroup) error
 	Update(u *ServiceGroup) error
 }
